+++ cdk/jitsi/user_data.sh
@@ -39,8 +39,9 @@ insert_logging_config "web" 6 "/root/jitsi-docker-jitsi-meet/docker-compose.yml"
 insert_logging_config "jibri" 5 "/root/jitsi-docker-jitsi-meet/jibri.yml"
 insert_logging_config "jigasi" 6 "/root/jitsi-docker-jitsi-meet/jigasi.yml"
 insert_logging_config "etherpad" 6 "/root/jitsi-docker-jitsi-meet/etherpad.yml"
+insert_logging_config "transcriber" 5 "/root/jitsi-docker-jitsi-meet/transcriber.yml"
 
-echo 's3fs#${AssetsBucket} /s3 fuse _netdev,allow_other,nonempty,iam_role=auto 0 0' >> /etc/fstab
+echo 's3fs#${AssetsBucket} /s3 fuse _netdev,allow_other,nonempty,iam_role=${IamRole} 0 0' >> /etc/fstab
 rm -rf /s3 && mkdir /s3
 mount -a
 mkdir -p /s3/jitsi-meet-cfg/{web,transcripts,prosody/config,prosody/prosody-plugins-custom,jicofo,jvb,jigasi,jibri}
@@ -53,7 +54,7 @@ ip_array=($ips)
 ip1=${!ip_array[0]}
 ip2=${!ip_array[1]}
 
-/root/check-secrets.py ${AWS::Region} ${SecretArn} ${EnableRecording} ${EnableEtherpad}
+/root/check-secrets.py ${AWS::Region} ${SecretArn}
 
 mkdir -p /opt/oe/patterns
 SECRET_NAME=$(aws secretsmanager describe-secret --secret-id ${SecretArn} | jq -r .Name)
@@ -63,6 +64,21 @@ aws ssm get-parameter \
     --query Parameter.Value \
 | jq -r . > /opt/oe/patterns/instance.json
 
+JICOFO_AUTH_PASSWORD=$(cat /opt/oe/patterns/instance.json | jq -r .JICOFO_AUTH_PASSWORD)
+JVB_AUTH_PASSWORD=$(cat /opt/oe/patterns/instance.json | jq -r .JVB_AUTH_PASSWORD)
+JIGASI_XMPP_PASSWORD=$(cat /opt/oe/patterns/instance.json | jq -r .JIGASI_XMPP_PASSWORD)
+JIBRI_RECORDER_PASSWORD=$(cat /opt/oe/patterns/instance.json | jq -r .JIBRI_RECORDER_PASSWORD)
+JIBRI_XMPP_PASSWORD=$(cat /opt/oe/patterns/instance.json | jq -r .JIBRI_XMPP_PASSWORD)
+JIGASI_TRANSCRIBER_PASSWORD=$(cat /opt/oe/patterns/instance.json | jq -r .JIGASI_TRANSCRIBER_PASSWORD)
+
+JITSI_IMAGE_VERSION=$(cat /root/jitsi-image-version)
+# custom .env
+CUSTOM_DOT_ENV_CONFIG="# no custom config defined"
+if [[ "${CustomDotEnvParameterArn}" != "" ]]; then
+    CUSTOM_DOT_ENV_CONFIG_TITLE="# custom config fetched from ${CustomDotEnvParameterArn}"
+    CUSTOM_DOT_ENV_CONFIG_VALUE=$(aws ssm get-parameter --name "${CustomDotEnvParameterArn}" --with-decryption --output text --query Parameter.Value)
+    CUSTOM_DOT_ENV_CONFIG=$(printf "%s\n\n%s" "$CUSTOM_DOT_ENV_CONFIG_TITLE" "$CUSTOM_DOT_ENV_CONFIG_VALUE")
+fi
 cat <<EOF > /root/jitsi-docker-jitsi-meet/.env
 CONFIG=/s3/jitsi-meet-cfg
 HTTP_PORT=80
@@ -71,6 +87,16 @@ TZ=UTC
 PUBLIC_URL=https://${Hostname}
 JVB_ADVERTISE_IPS=$ip1,$ip2
 ENABLE_LETSENCRYPT=0
+JITSI_IMAGE_VERSION=$JITSI_IMAGE_VERSION
+
+JICOFO_AUTH_PASSWORD=$JICOFO_AUTH_PASSWORD
+JVB_AUTH_PASSWORD=$JVB_AUTH_PASSWORD
+JIGASI_XMPP_PASSWORD=$JIGASI_XMPP_PASSWORD
+JIGASI_TRANSCRIBER_PASSWORD=$JIGASI_TRANSCRIBER_PASSWORD
+JIBRI_RECORDER_PASSWORD=$JIBRI_RECORDER_PASSWORD
+JIBRI_XMPP_PASSWORD=$JIBRI_XMPP_PASSWORD
+
+$CUSTOM_DOT_ENV_CONFIG
 EOF
 
 cat <<EOF > /root/start.sh
@@ -80,11 +106,14 @@ DOCKER_FILES="-f docker-compose.yml"
 if grep -q '^ENABLE_RECORDING=1' .env; then
   DOCKER_FILES="\$DOCKER_FILES -f jibri.yml"
 fi
-if grep -q '^JIGASI_SIP_URI=' .env; then
-  DOCKER_FILES="\$DOCKER_FILES -f jigasi.yml"
+if grep -q '^ENABLE_TRANSCRIPTIONS=1' .env; then
+  DOCKER_FILES="\$DOCKER_FILES -f transcriber.yml"
 fi
 if grep -q '^ETHERPAD_URL_BASE=' .env; then
   DOCKER_FILES="\$DOCKER_FILES -f etherpad.yml"
+fi
+if grep -q '^JIGASI_SIP_URI=' .env; then
+  DOCKER_FILES="\$DOCKER_FILES -f jigasi.yml"
 fi
 docker compose \$DOCKER_FILES up -d
 EOF
@@ -117,10 +146,22 @@ WorkingDirectory=/root/jitsi-docker-jitsi-meet
 [Install]
 WantedBy=multi-user.target
 EOF
-
-rm -f /s3/jitsi-meet-cfg/web/custom-config.js
-rm -f /s3/jitsi-meet-cfg/web/custom-interface_config.js
-/root/append-config.py
+# custom-config.js
+CUSTOM_CONFIG_JS_CONFIG="// no custom config defined"
+if [[ "${CustomConfigJsParameterArn}" != "" ]]; then
+    CUSTOM_CONFIG_JS_CONFIG_TITLE="// custom config fetched from ${CustomConfigJsParameterArn}"
+    CUSTOM_CONFIG_JS_CONFIG_VALUE=$(aws ssm get-parameter --name "${CustomConfigJsParameterArn}" --with-decryption --output text --query Parameter.Value)
+    CUSTOM_CONFIG_JS_CONFIG=$(printf "%s\n\n%s" "$CUSTOM_CONFIG_JS_CONFIG_TITLE" "$CUSTOM_CONFIG_JS_CONFIG_VALUE")
+fi
+# custom-interface_config.js
+CUSTOM_INTERFACE_CONFIG_JS_CONFIG="// no custom config defined"
+if [[ "${CustomInterfaceConfigJsParameterArn}" != "" ]]; then
+    CUSTOM_INTERFACE_CONFIG_JS_CONFIG_TITLE="// custom config fetched from ${CustomInterfaceConfigJsParameterArn}"
+    CUSTOM_INTERFACE_CONFIG_JS_CONFIG_VALUE=$(aws ssm get-parameter --name "${CustomInterfaceConfigJsParameterArn}" --with-decryption --output text --query Parameter.Value)
+    CUSTOM_INTERFACE_CONFIG_JS_CONFIG=$(printf "%s\n\n%s" "$CUSTOM_INTERFACE_CONFIG_JS_CONFIG_TITLE" "$CUSTOM_INTERFACE_CONFIG_JS_CONFIG_VALUE")
+fi
+echo "$CUSTOM_CONFIG_JS_CONFIG" > /s3/jitsi-meet-cfg/web/custom-config.js
+echo "$CUSTOM_INTERFACE_CONFIG_JS_CONFIG" > /s3/jitsi-meet-cfg/web/custom-interface_config.js
 
 systemctl enable jitsi.service
 systemctl start jitsi.service