+++ packer/ubuntu_2204_appinstall.sh
@@ -114,10 +114,11 @@ EOF
 # Jitsi configuration
 #
 
-# https://github.com/jitsi/docker-jitsi-meet/releases/tag/stable-9779 10/22/2024
-JITSI_VERSION=stable-9779
+# https://github.com/jitsi/docker-jitsi-meet/releases/tag/stable-9823 # 11/12/2024
+JITSI_VERSION=stable-9823
 
 cd /root
+echo $JITSI_VERSION >> /root/jitsi-image-version
 
 # s3fs
 apt-get update && apt-get -y install s3fs
@@ -145,8 +146,6 @@ import uuid
 
 region_name = sys.argv[1]
 arn = sys.argv[2]
-enable_recording = sys.argv[3]
-enable_etherpad = sys.argv[4]
 
 client = boto3.client("secretsmanager", region_name=region_name)
 response = client.get_secret_value(
@@ -162,30 +161,19 @@ if 'username' in current_secret:
     needs_update = True
     del current_secret['username']
 NEEDED_SECRETS_WITH_SIMILAR_REQUIREMENTS = [
-    ".env:JICOFO_AUTH_PASSWORD",
-    ".env:JVB_AUTH_PASSWORD",
-    ".env:JIGASI_XMPP_PASSWORD",
-    ".env:JIBRI_RECORDER_PASSWORD",
-    ".env:JIBRI_XMPP_PASSWORD"
+    "JICOFO_AUTH_PASSWORD",
+    "JVB_AUTH_PASSWORD",
+    "JIGASI_XMPP_PASSWORD",
+    "JIGASI_TRANSCRIBER_PASSWORD",
+    "JIBRI_RECORDER_PASSWORD",
+    "JIBRI_XMPP_PASSWORD"
 ]
 for secret in NEEDED_SECRETS_WITH_SIMILAR_REQUIREMENTS:
   if not secret in current_secret:
     needs_update = True
-    cmd = "random_value=\$(seed=\$(date +%s%N); tr -dc '[:alnum:]' < /dev/urandom | head -c 32; echo \$seed | sha256sum | awk '{print substr(\$1, 1, 32)}'); echo \$random_value"
+    cmd = "random_value=\$(seed=\$(date +%s%N); tr -dc '[:alnum:]' < /dev/urandom | head -c 16; echo \$seed | sha256sum | awk '{print substr(\$1, 1, 16)}'); echo \$random_value"
     output = subprocess.run(cmd, stdout=subprocess.PIPE, shell=True).stdout.decode('utf-8').strip()
     current_secret[secret] = output
-if enable_recording == 'true' and current_secret.get('.env:ENABLE_RECORDING') != '1':
-  needs_update = True
-  current_secret['.env:ENABLE_RECORDING'] = '1'
-if enable_recording == 'false' and current_secret.get('.env:ENABLE_RECORDING') == '1':
-  needs_update = True
-  del current_secret['.env:ENABLE_RECORDING']
-if enable_etherpad == 'true' and current_secret.get('.env:ETHERPAD_URL_BASE') != 'http://etherpad.meet.jitsi:9001':
-  needs_update = True
-  current_secret['.env:ETHERPAD_URL_BASE'] = 'http://etherpad.meet.jitsi:9001'
-if enable_etherpad == 'false' and current_secret.get('.env:ETHERPAD_URL_BASE') == 'http://etherpad.meet.jitsi:9001':
-  needs_update = True
-  del current_secret['.env:ETHERPAD_URL_BASE']
 if needs_update:
   client.update_secret(
     SecretId=arn,
@@ -196,45 +184,6 @@ else:
 EOF
 chown root:root /root/check-secrets.py
 chmod 744 /root/check-secrets.py
-
-cat <<EOF > /root/append-config.py
-#!/usr/bin/env python3
-
-import json
-import os
-with open('/opt/oe/patterns/instance.json', 'r') as file:
-    data = json.load(file)
-for key, value in data.items():
-    try:
-        file_path, varname = key.split(':')
-        # Check if file_path is a .js file
-        if os.path.splitext(file_path)[1] == '.js':
-            if file_path == 'interface_config.js':
-                output = f'interfaceConfig.{varname} = "{value}";'
-            elif file_path == 'config.js':
-                output = f'config.{varname} = "{value}";'
-            else:
-                output = f'{varname} = "{value}";'
-            file_path = f'/s3/jitsi-meet-cfg/web/custom-{file_path}'
-        else:
-            output = f'{varname}={value}'
-            file_path = f'/root/jitsi-docker-jitsi-meet/{file_path}'
-        # Ensure the directory exists
-        directory = os.path.dirname(file_path)
-        if directory and not os.path.exists(directory):
-            os.makedirs(directory)
-        # Check if file exists, if not create it
-        if not os.path.exists(file_path):
-            with open(file_path, 'w') as f:
-                pass  # Just create the file
-        # Append the output to the file
-        with open(file_path, 'a') as f:
-            f.write(output + '\n')
-    except Exception as e:
-        print(f'Error processing key {key}: {e}')
-EOF
-chown root:root /root/append-config.py
-chmod 744 /root/append-config.py
 
 # post install steps
 curl -O "https://raw.githubusercontent.com/ordinaryexperts/aws-marketplace-utilities/$SCRIPT_VERSION/packer_provisioning_scripts/$SCRIPT_POSTINSTALL"